feat(table): rewrite cell components
