+++ frontend/src/components/database/AppSchema.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { useRouter } from "vue-router";
 import { useUrlParams } from "../../composables/useUrlParams";
-import { FormattedQueryResult, RowAction } from "./table";
+import { FormattedQueryResult, RowAction } from "./table/table";
 import { useWails } from "../../wails";
 import { GetSchemaTables } from "../../../wailsjs/go/app/App";
 import { Effect } from "effect";
