+++ frontend/src/components/database/table/AppRows.vue
